requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pyyaml==6.0.2
transformers==4.41.2
accelerate==0.33.0
sentence-transformers==3.0.1
torch>=2.2.0
numpy>=1.24
scikit-learn==1.4.2
pypdf==4.2.0
fpdf2==2.7.9
ollama==0.3.3
